export.m5() can treate variable number (or no) orientation points from whole coordinate list
In sf_zartkert.R now the whole coordinate list sent to the m5 processing.

In the export.m5() now orientation is filtered and sp-s before
orientation added. For generalisation first and last sp-s are also
substracted. Because overlapping duplication in rownumbers are
eliminated with unique.

This commit resolves #8 issue. Multiple orietation from one point is not tested.