Correct instrument hight is calculated in the case of topo points
The topo calculation is divided and traverse instrument height generation
is inserted between two.

In the topo processing the stations with topo point are selected and
height corrected and additionally correct instrument height re-calculated.

This commit resolves #24 and closes #19 because it replace the necessity
to store instrument heights in some additional datasets.